Fix MSVC warning C4267: size_t to int conversion in symmetro
Added explicit static_cast<int>(i) when passing loop index to
setMult() to silence MSVC warning on line 1284.

Warning was:
C4267: 'argument': conversion from 'size_t' to 'const int',
possible loss of data

The cast is safe here because the loop index is always 0 or 1
(at most 2 iterations), well within int range.

+++ base/antiprism_c_api.cc
@@ -1281,7 +1281,7 @@ ANTIPRISM_API AntiStatus anti_make_symmetro_kaplan_hart(
 
     // Set multipliers using the non-zero indices
     for (size_t i = 0; i < idx.size(); i++) {
-      symm.setMult(i, mults[idx[i]]);
+      symm.setMult(static_cast<int>(i), mults[idx[i]]);
     }
 
     // Fill symmetry vectors (Kaplan-Hart mode)
